Implement <my_engine>:assets.
It copies assets from the gem to the host app.  An initializer is
added to lib/<my_engine>/engine.rb, setting paths.public to the
host app's paths.public/<my_engine>. The original engine's
paths.public is saved in paths.gem_public, which is used in the rake
task to find the original assets.