Regearing asset stuff, as the original strategy won't work.
To separate engine assets from host assets, we're going to exploit
ActionController::Base.config.asset_path.  This config property needs
to be set on each engine controller.

It is, however, inheritable.  It would be enough to set it once on
the engine's ApplicationController superclass.  The catch is that,
out of the box, ApplicationController is not name-spaced.  So it is
shared (reopened) by all engines and the host app, which defeats us.

The engine author almost certainly wants her models, controllers, etc
to be name-spaced in one way or another, so engineer:install does just
enough to get by:  it creates

app/controllers/<my_engine>/application_controller.rb

..by copying the previous application_controller.rb into it and
tweaking the class' module.  The old application_controller.rb is
removed and all other controllers will switch their superclass from
< ApplicationController to < MyEngine::ApplicationController as
needed.  The other controllers are not moved or otherwise changed;
the engine author can do that when she's ready.

This does not yet actually set the asset_path from the engine class,
there's some homework to do about the timing of when to do that.