Rejigger asset_path stuff once more.
Setting asset_path to "/my_engine%s" on engine controllers is good
enough when there is no mixing of views.

When the engine wants to render with the host application's layout,
it creates problems. The host layout renders from the engine
controller, and so all asset paths in the layout point into the
engine.  The least-surprise behavior would have the host layout
render with the host's asset_path.

We address this by monkey-patching ActionView::Template#render.
On each invocation, the template's identifier is pushed onto a
thread-local stack (and popped on exit.)  This identifier is the
file system path of the template for fs-based templates, and so can
identify the template's owning engine in that case.

The engine asset_path becomes a lambda which examines the latest
view identifier.  The engine asset path is used if the view is not a
file system path owned by the host application.  Thus the default
case favors the engine path for non-fs templates.