Implement <my_engine>:db:schema, which imports an engine's db/schema.rb as a migration.
This presents some challenges, since a "schema migration" accumulates the other, regular
migrations that eventually generated it.  <my_engine>:db:migrate detects and skips these "implied"
migrations with a naming convention: the schema migration embeds the name of the last migration it
implies.  So, it is named something like my_engine_schema_after_create_posts.

When there are no migrations to skip (say, because the engine author removed them), this shortens
to my_engine_schema.